sys: mem: Fix mutex issues in malloc
The entering the mutex happens too early. Atomicity is not required
when checking wheather or not the requested memory exceeds the ammount
of system memory, as there are no changing variables in there. This
could potentially cause a delay where non is necessary.

Also in the allocation loop itself, there is a return NULL statement
which is called if a certain variable is NULL. This is done without
releasing the mutex, causing a self induced deadlock in any future
calls to malloc.

Both issues are now fixed.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>